requirements-basic.txt
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
pymongo>=4.0.0
elasticsearch>=8.0.0
requests>=2.25.0
python-dotenv>=0.19.0
folium>=0.12.0
tqdm>=4.60.0
geopy>=2.0.0